Update vision inspection tool
