+++ squid_control/start_hypha_service.py
@@ -4304,7 +4304,6 @@ class MicroscopeHyphaService:
                 object_uuid=target_uuid,
                 application_id=actual_experiment_name,
                 limit=limit_expected,
-                base_url="https://hypha.aicell.io/agent-lens/apps/agent-lens"
             )
             
             if not search_result.get('success'):
